Extends isProtectedRoute to allow HEAD requests
This commit is to address a problem that has been perplexing me for
months. Whenever I try to make a HEAD request to a clay page I get a 302
redirect to `/_auth/login` instead of getting response headers from the
page.

The consequence of this behavior is that one is forced to render and
download the full page contents, when they may only be interested in the
Content-Length, status code, or other bits of information.