introduces per-user API tokens via 'apikey: true'
This commit introduces a rudimentary approach to assigning API keys on
a per-user basis (rather than one global plain-text environment
variable).

**Generating an API Key**

When making a `POST` request to the `/_users/` endpoint the inclusion
of the following in the request body will generate an API key for the
user:

```
POST /_users/

{
  "apikey": true,
  "auth": "admin",
  "provider": "google",
  "username": "test@example.com"
}
```

The response will look like this. _Note: This is the only opportunity
to retrieve the generated key. Keys are hashed in the database._

```js
{
  "_ref": "/_users/dGVzdEBleGFtcGxlLmNvbUBnb29nbGU=",
  "apikey": "dGVzdEBleGFtcGxlLmNvbUBnb29nbGU=:Z8EVBDN-M1T4EED-NH93XNJ-8J1374R",
  "auth": "admin",
  "provider": "google",
  "username": "test@example.com"
}
```

**Revoking an API Key**

To remove a user's key `POST` again with `apikey: false`:

```
POST /_users/

{
  "apikey": false,
  "auth": "admin",
  "provider": "google",
  "username": "test@example.com"
}
```

The response will look like this.

```js
{
  "_ref": "/_users/dGVzdEBleGFtcGxlLmNvbUBnb29nbGU=",
  "apikey": null,
  "auth": "admin",
  "provider": "google",
  "username": "test@example.com"
}
```

---

In addition to per-user API keys, this commit introduces a new
environment variable `CLAY_DISABLE_GLOBAL_ACCESS_KEY`. When set to a
"true" value (`t`, `true`, `1`) _only_ user-level keys will work.

This provides a "single-user-mode"-esq toggle for emergency recovery,
allows Clay sites to run in a secure mode, but maintains backward
compatibility for old API environment-based API keys where required or
where convenient.